chore: separate core classes from actual feature implementations
