+++ src/features/accessibility.py
@@ -4,8 +4,8 @@ import json
 from aiohttp import ClientConnectorError, ClientSession
 
 from app.models import Explanation, StarCase
-from features.metadata_base import MetadataBase
-from features.website_manager import WebsiteData
+from core.metadata_base import MetadataBase
+from core.website_manager import WebsiteData
 from lib.constants import (
     ACCESSIBILITY,
     DESKTOP,
