+++ src/features/gdpr.py
@@ -1,8 +1,8 @@
 import re
 
 from app.models import Explanation, StarCase
-from features.metadata_base import MetadataBase
-from features.website_manager import WebsiteData
+from core.metadata_base import MetadataBase
+from core.website_manager import WebsiteData
 from lib.constants import STRICT_TRANSPORT_SECURITY, VALUES
