+++ src/features/malicious_extensions.py
@@ -1,6 +1,6 @@
 from app.models import Explanation, StarCase
-from features.metadata_base import MetadataBase
-from features.website_manager import WebsiteData
+from core.metadata_base import MetadataBase
+from core.website_manager import WebsiteData
 from lib.constants import VALUES
