+++ tests/integration/security_test.py
@@ -1,8 +1,8 @@
 import json
 
 from app.models import Explanation, StarCase
+from core.website_manager import WebsiteData
 from features.security import Security
-from features.website_manager import WebsiteData
 from lib.logger import get_logger
 from tests.integration.features_integration_test import _test_feature
