+++ tests/unit/metadata_manager_test.py
@@ -7,9 +7,9 @@ import pytest
 
 from app.models import Explanation, StarCase
 from cache.cache_manager import CacheManager
+from core.metadata_manager import MetadataManager
 from features.html_based import Advertisement
 from features.javascript import Javascript
-from features.metadata_manager import MetadataManager
 from lib.constants import (
     ACCESSIBILITY,
     EXPLANATION,
@@ -87,7 +87,7 @@ def test_cache_data(metadata_manager: MetadataManager):
     cache_manager = get_cache_manager()
 
     with mock.patch(
-        "features.metadata_manager.create_cache_entry"
+        "core.metadata_manager.create_cache_entry"
     ) as create_cache_entry:
         metadata_manager.cache_data(meta_data, cache_manager, allow_list)
         assert create_cache_entry.call_args[0][2][VALUES] == []
@@ -115,7 +115,7 @@ def test_extract_meta_data(metadata_manager: MetadataManager):
     cache_manager.domain = test_host
 
     with mock.patch(
-        "features.metadata_manager.shared_memory.ShareableList"
+        "core.metadata_manager.shared_memory.ShareableList"
     ) as shareable_list:
         with mock.patch(
             "cache.cache_manager.CacheManager._class.is_enough_cached_data_present"
@@ -142,7 +142,7 @@ def test_extract_meta_data(metadata_manager: MetadataManager):
     cache_manager.domain = test_host
 
     with mock.patch(
-        "features.metadata_manager.shared_memory.ShareableList"
+        "core.metadata_manager.shared_memory.ShareableList"
     ) as shareable_list:
         with mock.patch(
             "cache.cache_manager.CacheManager._class.is_enough_cached_data_present"
@@ -195,24 +195,24 @@ def test_start(mocker, metadata_manager: MetadataManager):
     log_spy = mocker.spy(metadata_manager, "_logger")
 
     with mock.patch(
-        "features.metadata_manager.shared_memory.ShareableList"
+        "core.metadata_manager.shared_memory.ShareableList"
     ) as shareable_list:
         with mock.patch(
-            "features.metadata_manager.WebsiteManager"
+            "core.metadata_manager.WebsiteManager"
         ) as website_manager:
             with mock.patch(
                 "cache.cache_manager.CacheManager._class.update_to_current_domain"
             ):
                 async_mock = AsyncMock(return_value={})
                 with mock.patch(
-                    "features.metadata_manager.MetadataManager._class._extract_meta_data",
+                    "core.metadata_manager.MetadataManager._class._extract_meta_data",
                     side_effect=async_mock,
                 ) as extract_meta_data:
                     with mock.patch(
-                        "features.metadata_manager.MetadataManager._class.cache_data"
+                        "core.metadata_manager.MetadataManager._class.cache_data"
                     ):
                         with mock.patch(
-                            "features.metadata_manager.WebsiteManager._class.load_website_data"
+                            "core.metadata_manager.WebsiteManager._class.load_website_data"
                         ):
                             website_manager.get_instance().website_data.html = (
                                 "non-empty html"
